Fix bug in "visibility" calculation
The problem is that the previous implementation assumed incorrectly that
the two nearest points would never be on the same "side" of the
observation point. The new implementation doesn't bother with all that.
It splits the points on the line to groups on the left and right
(x-wise) of the observation point and uses that as the "near points"
count. Clever? We'll see how it holds up in part 2 :D